Add the script to automate creating the Amazon Linux OVF file
Amazon provides a disk image and expects you to create an ISO with cloud-init files, which will be read when the disk boots. This automates creating the ISO, creating the VM from the disk and exporting the OVF file. Next step is consuming that file in a packer template and then we have a system.

Signed-off-by: Tim Smith <tsmith@chef.io>