Switch to shelling out for all vagrant uploads
The vagrant cloud gem is pretty fundamentally broken. It does not allow uploading a box that is not directly owned by the user performing the upload. Since I'm not using the bento account directly, and that seems like a horrible idea, this won't work. Instead we can just shell out to the vagrant CLI to perform the uploads. The vagrant CLI also does a much better job of this.

Other fixes:

Simplifies how we handle slugs by just using a match on the name. This means we don't have to update the slug list with the exact versions to match on every time we bump the version.

Signed-off-by: Tim Smith <tsmith@chef.io>